refactor: rename Django app from feed to engine
