+++ core/engine/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class FeedConfig(AppConfig):
+class EngineConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "feed"
+    name = "engine"
