+++ core/infobus/urls.py
@@ -31,6 +31,6 @@ urlpatterns = [
     path("", include("website.urls")),
     path("api/", include("api.urls")),
     path("gtfs/", include("gtfs.urls")),
-    path("status/", include("feed.urls")),
+    path("status/", include("engine.urls")),
     path("alertas/", include("alerts.urls")),
 ]
